pyproject.toml
[tool.poetry]
name = "chess-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
chromadb = "^0.5.18"
openai = "^1.54.4"
llama-index = "0.10.50"
python-dotenv = "^1.0.1"
streamlit = "^1.40.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
